userAuth done for front end

+++ frontend/src/components/WorkoutDetails.js
@@ -1,13 +1,19 @@
 import { useWorkoutContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) =>{
   const { dispatch } = useWorkoutContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if(!user){
+      return
+    }
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {'Authorization': `Bearer ${user.token}`}
     })
     const json = await response.json()
